Make SaaSquatchClient an interface
